Removed the Cluster 2 WBD Overview Postscript Display Data Description
The Cluster Active Archive (CAA) provides PNG format versions of the Overview plots. The University of Iowa provides Postscript format. The switch to using CAA over Iowa meant a switch in these formats. The old description was removed in favor of this newer version. More details on other differences are provided in the previous message.